+++ CHANGELOG.md
@@ -55,6 +55,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - [2299](https://github.com/umee-network/umee/pull/2299) Upgrade Cosmos SDK to v0.47.
 - [2301](https://github.com/umee-network/umee/pull/2301) use gov/v1 MinInitialDepositRatio and set it to 0.1.
 
+### Breaking Changes
+
+- [2332](https://github.com/umee-network/umee/pull/2332) Rename: `TotalOutflowSum` to `OutflowSum`, `TotalInflowSum` to `InflowSum`.
+
 ## v6.1.0 - 2023-10-17
 
 ### Improvements